requirements.txt
Flask==3.1.0
Flask-Session==0.8.0
numpy==2.1.3
PyYAML==6.0.2
matplotlib==3.9.3
pandas==2.2.3
seaborn==0.13.2

